pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Multi-agent system built with deepagents framework, featuring orchestrator-driven task routing and extensible sub-agent architecture."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.6",
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.3",
    "langchain-google-genai>=3.1.0",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
]
langgraph = [
    "langgraph-cli[inmem]>=0.1.55",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agents", "subagents", "tools", "skills"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "D", "UP"]
ignore = ["UP006", "UP007"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
